rewite calibration flow
